Add syslog_ip_address variable
Add separate variable instead of use default ipv4 address. There are
cases when server won't have any public ipv4 - e.g. if it will be hosted
on AWS/Azure/Openstack. So add an ability to override ip for which
certificate will be issued and cert file will be named upon.